@@ -3,18 +3,18 @@
     "description": "A package of useful components for internationalisation in Laravel applications.",
     "type": "library",
     "require": {
-        "php": ">=7.1",
+        "php": ">=7.2",
         "ext-intl": "*",
-        "illuminate/config": "5.*",
-        "illuminate/support": "5.*",
-        "illuminate/view": "5.*",
+        "illuminate/config": "^6.0",
+        "illuminate/support": "^6.0",
+        "illuminate/view": "^6.0",
         "league/iso3166": "^2.1",
-        "guzzlehttp/guzzle": "^6.3",
+        "guzzlehttp/guzzle": "^7.0",
         "mpociot/vat-calculator": "^2.4"
     },
     "require-dev": {
-        "satooshi/php-coveralls": "^1.0",
-        "phpunit/phpunit": "^7.5"
+        "php-coveralls/php-coveralls": "^2.0",
+        "phpunit/phpunit": "^8.5"
     },
     "license": "MIT",
     "authors": [
