@@ -21,7 +21,7 @@ class CountryLocaleViewComposerTest extends TestCase
         /** @var View|\PHPUnit_Framework_MockObject_MockObject $viewMock */
         $viewMock = $this->getMockBuilder(View::class)
             ->disableOriginalConstructor()
-            ->setMethods(['with'])
+            ->onlyMethods(['with'])
             ->getMock();
 
         $viewMock->expects($this->once())
